requirements.txt
Flask==0.12.2
google-cloud==0.19.0
BeautifulSoup4==4.6.0